feat : 게시물 등록, 수정, 삭제에 로그인 확인 기능 개발

+++ services/posting.py
@@ -7,7 +7,7 @@ def get_posting_preview_list(bunch: int, db: sql.cursors.DictCursor) -> List[R_P
         "SELECT posting_id, posting_title, posting_header_image_id, posting_preview, posting_datetime, comment_count, like_count, user_id, user_profile_image_id " \
         "FROM posting " \
         "NATURAL JOIN users " \
-        "ORDER BY posting_datetime " \
+        "ORDER BY posting_datetime DESC " \
         "LIMIT 20 OFFSET %s;",
         bunch * 20
     )
@@ -17,17 +17,18 @@ def get_posting_preview_list(bunch: int, db: sql.cursors.DictCursor) -> List[R_P
 
     return result
 
-def register_posting(posting: R_RegistingPosting, db: sql.cursors.DictCursor) -> int | None:
+def register_posting(posting: R_RegistingPosting, user_id: int, db: sql.cursors.DictCursor) -> int | None:
     
     try: 
         db.execute(
-            "INSERT INTO posting (posting_title, posting_header_image_id, posting_preview, content) " \
-            "VALUES (%s, %s, %s, %s)",
+            "INSERT INTO posting (posting_title, posting_header_image_id, posting_preview, content, user_id) " \
+            "VALUES (%s, %s, %s, %s, %s)",
             (
                 posting.posting_title,
                 posting.posting_header_image_id,
                 posting.posting_preview,
-                posting.content
+                posting.content,
+                user_id
             )
         )
     except sql.err.IntegrityError:
@@ -52,19 +53,20 @@ def get_posting(posting_id: int, db: sql.cursors.DictCursor) -> R_Posting | None
 
     return result
     
-def change_posting(posting_id: int, put_data: R_RegistingPosting, db: sql.cursors.DictCursor) -> None:
+def change_posting(posting_id: int, put_data: R_RegistingPosting, user_id:int, db: sql.cursors.DictCursor) -> None:
     db.execute(
         "UPDATE posting " \
         "SET posting_title = %s, posting_header_image_id = %s, posting_preview = %s, content = %s " \
-        "WHERE posting_id = %s;",
+        "WHERE posting_id = %s AND user_id = %s;",
         (
             put_data.posting_title,
             put_data.posting_header_image_id,
             put_data.posting_preview,
             put_data.content,
-            posting_id
+            posting_id,
+            user_id
         )
     )
 
-def remove_posting(posting_id: int, db: sql.cursors.DictCursor) -> None:
-    db.execute("DELETE FROM posting WHERE posting_id = %s", posting_id)
+def remove_posting(posting_id: int, user_id: int, db: sql.cursors.DictCursor) -> None:
+    db.execute("DELETE FROM posting WHERE posting_id = %s AND user_id = %s;", (posting_id, user_id))
